delete refactor
I'd probably do something more like this.

A single delete function, and then separate behavior around how to map
the delete to the selection and then update the queryCache.

To me, that separates the concerns a bit more clearly.

There's probably more to do here, but I'm a bit limited on time.